feat: add react-query to my created courses page
