--- src/query-data.ts
@@ -1,4 +1,0 @@
-const queryKeys = {
-  coursesList: ['allCourses'],
-  CourseDetailsPage: ['courseDetails', 'courseId'],
-};
