+++ src/react-query-data.ts
@@ -1,0 +1,6 @@
+// Created to know which keys are used in a react-query library application
+const queryKeys = {
+  CoursesList: ['allCourses'],
+  CourseDetailsPage: ['courseDetails', 'courseId'],
+  MyCreatedCoursesPage: ['myCreatedCourses', 'initDataRaw'],
+};
